% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_baseline}
\alias{CAP_baseline}
\title{Correct Attribution Probability Baseline}
\usage{
CAP_baseline(dataframe, target_variable)
}
\arguments{
\item{dataframe}{dataset}
}
\description{
This function allows you to determine the baseline Correct Attribution Probability (CAP) of a given dataset.
}
