% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_baseline_num}
\alias{CAP_baseline_num}
\title{Correct Attribution Probability Baseline(Numeric)}
\usage{
CAP_baseline_num(dataframe, target_variable, y = 2500)
}
\arguments{
\item{dataframe}{dataset}

\item{y}{Continuous numerical values of the target variable are rounded to the nearest value of y. Default y = 2500.}
}
\description{
This function allows you to determine the baseline Correct Attribution Probability (CAP) of a given dataset.
}
