% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_original}
\alias{CAP_original}
\title{Correct Attribution Probability Original}
\usage{
CAP_original(dataframe, key_variable, target_variable)
}
\arguments{
\item{dataframe}{dataset}

\item{key_variable}{key variable}

\item{target_variable}{target_variable}
}
\description{
This function allows you to determine the Correct Attribution Probability (CAP) of a given dataset.
}
