% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_original_num}
\alias{CAP_original_num}
\title{Correct Attribution Probability Original (numeric)}
\usage{
CAP_original_num(dataframe, key_variable, target_variable, y = 2500)
}
\arguments{
\item{dataframe}{dataset}

\item{key_variable}{key variable}

\item{target_variable}{target_variable}

\item{y}{Continuous numerical values of the target variable are rounded to the nearest value of y. Default y = 2500.}
}
\description{
This function allows you to determine the Correct Attribution Probability (CAP) of a given dataset where the target variable is numeric.
}
