% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_synthetic}
\alias{CAP_synthetic}
\title{Correct Attribution Probability for Synthetic Dataset}
\usage{
CAP_synthetic(dataframe, synthetic_dataframe, key_variable, target_variable)
}
\arguments{
\item{dataframe}{dataset}

\item{synthetic_dataframe}{synthetic dataset}

\item{key_variable}{key variable}

\item{target_variable}{target variable}
}
\description{
This function allows you to determine the baseline Correct Attribution Probability (CAP) of a synthetic dataset, relative to an original dataset.
}
