% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAP_functions.R
\name{CAP_synthetic_num}
\alias{CAP_synthetic_num}
\title{Correct Attribution Probability for Synthetic Dataset (Numeric)}
\usage{
CAP_synthetic_num(
  dataframe,
  synthetic_dataframe,
  key_variable,
  target_variable,
  y = 2500
)
}
\arguments{
\item{dataframe}{dataset}

\item{synthetic_dataframe}{synthetic dataset}

\item{key_variable}{key variable}

\item{target_variable}{target variable}

\item{y}{Continuous numerical values of the target variable are rounded to the nearest value of y. Default y = 2500.}
}
\description{
This function allows you to determine the baseline Correct Attribution Probability (CAP) of a synthetic dataset, relative to an original dataset. The target variable here is numeric.
}
