% File cinf-free-wilson.R
\name{make_freq_table}
\alias{make_freq_table}
\title{Make frequency table}
\description{
  Makes frequency table in Free-Wilson analysis
}
\usage{
make_freq_table(t, sample)
}
\arguments{
	\item{t}{initial table for Free-Wilson analysis}
	\item{sample}{generated sample}
}
\value{Frequency table}
\examples{
}
