% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMF.R
\name{p_gradUsparseWrapper}
\alias{p_gradUsparseWrapper}
\title{Internal function for computing the gradients}
\usage{
p_gradUsparseWrapper(r, par, stochastic = FALSE)
}
\arguments{
\item{r}{?}

\item{par}{?}

\item{stochastic}{Whether or not to perform updates on a subsample}
}
\value{
Gradient
}
\description{
Internal function for computing the gradients
}
