% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_functions.R
\name{ROC_indiv}
\alias{ROC_indiv}
\title{Ratio of Counts (Individual) Function}
\usage{
ROC_indiv(data1, data2, x, num_cat = 2000)
}
\arguments{
\item{data1}{1st dataset}

\item{data2}{2nd dataset}

\item{x}{column name that we want to compare ROC on}

\item{num_cat}{Number of categories that can be evaluated. Defaults to 2000}
}
\description{
This function allows you to determine the Ratio of Counts(ROC) for a particular variable between 2 datasets, whereby the
ratio of counts is an indication of the overlap between 2 datasets.
}
\keyword{Counts}
\keyword{ROC,}
\keyword{ROC_indiv()}
\keyword{Ratio}
\keyword{of}
