% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_functions.R
\name{ROC_list}
\alias{ROC_list}
\title{Ratio of Counts List Function}
\usage{
ROC_list(data1, data2, num_cat = 10000)
}
\arguments{
\item{data1}{1st dataset}

\item{data2}{2nd dataset}

\item{num_cat}{Number of categories that can be evaluated. Defaults to 2000}
}
\description{
This function allows you to determine the Ratio of Counts(ROC) for all variables between 2 datasets, whereby the
ratio of counts is an indication of the overlap between 2 datasets.
}
\keyword{Counts}
\keyword{ROC,}
\keyword{ROC_list()}
\keyword{Ratio}
\keyword{of}
