% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_functions.R
\name{ROC_numeric}
\alias{ROC_numeric}
\title{Ratio of Counts Numeric Function}
\usage{
ROC_numeric(data1, data2, x, y = 2500, num_cat = 2000)
}
\arguments{
\item{data1}{1st dataset}

\item{data2}{2nd dataset}

\item{x}{column name that we want to compare ROC on}

\item{y}{All values in y will be rounded off to the nearest y. Defaults to 2500. Should be adjusted according to the data structure.}

\item{num_cat}{Number of categories that can be evaluated. Defaults to 2000}
}
\description{
This function allows you to determine the Ratio of Counts(ROC) for a particular numeric variable between 2 datasets, whereby the
ratio of counts is an indication of the overlap between 2 datasets.
This function bins continuous numerical variables into discrete categories before calculating the ROC score.
}
\keyword{Counts}
\keyword{ROC,}
\keyword{ROC_numeric()}
\keyword{Ratio}
\keyword{of}
