% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_functions.R
\name{ROC_score}
\alias{ROC_score}
\title{Ratio of Counts Function}
\usage{
ROC_score(data1, data2, num_cat = 2000, min_items = 5)
}
\arguments{
\item{data1}{1st dataset}

\item{data2}{2nd dataset}

\item{num_cat}{Number of categories that can be evaluated. Defaults to 2000}

\item{min_items}{Minimum mumber of items in a category. Defaults to 5. If the number falls below this threshold, this particular variable will be omitted from the calculation of the ROC score.}
}
\description{
This function allows you to determine the Ratio of Counts(ROC) between 2 datasets, whereby the
ratio of counts is an indication of the overlap between 2 datasets.
Specifically, it calculates the ROC between 2 datasets for each variable. The overall ROC is the arithmetic mean of all the ROC scores.
}
\keyword{Counts}
\keyword{ROC,}
\keyword{Ratio}
\keyword{of}
