% File cinf-free-wilson.R
\name{sample_subst}
\alias{sample_subst}
\title{Produces data frame with samples of substitutions}
\description{
  Produces data frame with samples of substitutions taken from table t from multinomial distributions
}
\usage{
sample_subst(t, ssize)
}
\arguments{
	\item{t}{input data frame containing table}
	\item{ssize}{sample size}
}
\value{Data frame with samples of substitutions taken from table t from multinomial distributions}
\examples{
}
